requirements.txt
snowflake-connector-python==3.10.0
